Added some checks for read-only mode.
From what I've read, Discourse won't run our job if read-only mode is
on, but I don't think that stops read-only mode turning on after the job
has started. Since the job can take a while, it now checks regularly and
bails out if read-only mode turns on. It'll still write a log entry to
the database (if it wrote the Start entry) to ensure the Start and End
entries are balanced, but will stop writing soon after.